requirements.txt
flask==3.0.2
ai21==2.0.3
python-dotenv==1.0.1
